+++ services/attester/standard/service.go
@@ -16,6 +16,7 @@ package standard
 import (
 	"bytes"
 	"context"
+	"encoding/json"
 	"fmt"
 	"sync"
 	"time"
@@ -147,7 +148,12 @@ func (s *Service) Attest(ctx context.Context, data interface{}) ([]*phase0.Attes
 		return nil, errors.Wrap(err, "failed to obtain attestation data")
 	}
 	attestationData := attestationDataResponse.Data
-	log.Trace().Dur("elapsed", time.Since(started)).Msg("Obtained attestation data")
+	if e := log.Trace(); e.Enabled() {
+		data, err := json.Marshal(attestationData)
+		if err == nil {
+			e.Dur("elapsed", time.Since(started)).RawJSON("data", data).Msg("Obtained attestation data")
+		}
+	}
 
 	if attestationData.Slot != duty.Slot() {
 		s.monitor.AttestationsCompleted(started, duty.Slot(), len(validatorIndices), "failed")