+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 dev:
   - add proposal value and blinded status to trace
+  - add beaconblockproposer.builder-boost-factor
 
 1.9.0-alpha.2:
   - allow Vouch to act as an MEV-boost client for non-Vouch validators